move uniswap routes to shared types
